add type cast

+++ jump.go
@@ -17,7 +17,7 @@ func Hash(key uint64, numBuckets int) int32 {
 
 	for j < int64(numBuckets) {
 		b = j
-		key = key*2862933555777941757 + 1
+		key = key* uint64(2862933555777941757) + 1
 		j = int64(float64(b+1) * (float64(int64(1)<<31) / float64((key>>33)+1)))
 	}
